Use TestFile from PackageInfo.g as fallback
... if GAP_TESTFILE is not set. Also handle .tst files correctly.

Note that right now `GAP_TESTFILE` has a default value `tst/testall.g`.
In the future, we should consider dropping this default value, and
packages which for some reason want to use a different testfile in CI
versus what is specified in `PackageInfo.g` must then set `GAP_TESTFILE`
explicitly. But I'd rather not do this at the same time as introducing
this change.