Actually honor GAP_TESTFILE, switch its default
By accident the value of the GAP_TESTFILE input was completely ignored. Nobody
noticed, however, because the only packages using it have set it to the
TestFile value in their PackageInfo.g anyway. So we may as well use this
opportunity to also change its default value to an empty string