Fix regression for autoloaded packages
Some packages are autoloaded by GAP during startup. For these, the call to
`SetPackagePath` may fail for a silly reason: GAP compares it to the path of
the currently loaded version, and if they differ, prints an error. We give GAP
`$PWD` as path of the package; but in practice, GAP already loaded it from
`/tmp/gaproot/pkg/PackageName`. So use that as path instead.